feat: Refactor home component for dynamic weather locations
The home component has been refactored to dynamically generate links based on an array of weather locations. The HTML files for the app and home components have been deleted, and their content moved directly into the TypeScript file of the home component. CommonModule was also imported to enable usage of Angular's common directives in the template.